We ended up with even weirder problems with rbenv.
Since we had originally used a system wide rvm install on the
previous server we felt it was best to do that, and it finally
resolved our problems.

With poise-ruby gem install pg was unable to find libpq-dev even
though it was installed.

With rbenv capistrano was dying while compiling assets saying it
could not find libruby2.0, limited time so we moved on.